+++ bin/orm.dart
@@ -1,5 +1,0 @@
-import 'package:jaguar_orm/generator/generator/generator.dart';
-
-main(List<String> args) {
-  start(args);
-}