+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.2.0
+
+- Removed statement generation
+- Moved generator to its own package
+
 ## 0.0.1
 
 - Initial version, created by Stagehand